Make eliminators use Neutral instead of Normal
There is a bit of art in deciding what should be what, but in general for each of your types you have terms for introduction (like lambda, succ, 0) and you have terms to eliminate (like application, isZero etc.). Introduction things produce things of the type, elimination consumes things of the type. Introduction should be Normal, with normal children. Elimination should be neutral and the term consumed should be neutral, other should be normal