Do an extra quote/eval for normalisation tests
Normalisation under binders happens during quoting, and `conv` works on
Values. So to run these tests while making sure things are fully
normalised the fuctions to run are:

```
eval -> quote -> eval -> conv
```

Which is a bit inefficient but it'll do for now until there's proper
CBOR representation for the normal form.

Signed-off-by: Alex Humphreys <alex.humphreys@here.com>